add command for previewing table contents
