+++ extensions/positron-connections/package.nls.json
@@ -2,5 +2,7 @@
 	"displayName": "Positron Connections",
 	"description": "Provides a view of active database connections.",
 	"view.description": "Database Connections",
-	"view.welcome": "No database connections are currently active."
+	"view.welcome": "No database connections are currently active.",
+	"commands.previewTable.title": "Preview Table",
+	"commands.previewTable.category": "Connections"
 }
